Added reserved key mechanism to logging
Now, if the key ${ignore=default} is included (keys are put in between ${}) then the default hoorn log output ignores.

Useful to log certain things you don't want to clutter the output (for example in a CLI app) but still want written and don't want to remove the default hoornlog output for other things.